Fixed movement and did a SUPER BETA version of NPC's
As much as I didn't want to, I had to get rid of SDL_AddTimer. Since it
runs in a separate thread and does calculations at the same time as the
main thread, a bunch of the data was getting messed up during the loop,
so it was just easiest to remove it. (Just refresh the world a lot to
see a village house spawn)